move transport related parts to r_packet_transport from r_packet.py (which is only for definitions)
