+++ tardis/montecarlo/tests/test_montecarlo.py
@@ -50,6 +50,7 @@ import numpy as np
 import pandas as pd
 import tardis.montecarlo.montecarlo_numba.formal_integral as formal_integral
 import tardis.montecarlo.montecarlo_numba.r_packet as r_packet
+import tardis.montecarlo.montecarlo_numba.r_packet_transport
 import tardis.montecarlo.montecarlo_numba.utils as utils
 import tardis.montecarlo.montecarlo_configuration as mc
 from tardis import constants as const
@@ -445,7 +446,7 @@ def test_move_packet_across_shell_boundary_emitted(
     energy = 0.9
     packet = r_packet.RPacket(r, mu, nu, energy)
     packet.current_shell_id = current_shell_id
-    r_packet.move_packet_across_shell_boundary(
+    tardis.montecarlo.montecarlo_numba.r_packet_transport.move_packet_across_shell_boundary(
         packet, delta_shell, no_of_shells
     )
     assert packet.status == r_packet.PacketStatus.EMITTED
@@ -464,7 +465,7 @@ def test_move_packet_across_shell_boundary_reabsorbed(
     energy = 0.9
     packet = r_packet.RPacket(r, mu, nu, energy)
     packet.current_shell_id = current_shell_id
-    r_packet.move_packet_across_shell_boundary(
+    tardis.montecarlo.montecarlo_numba.r_packet_transport.move_packet_across_shell_boundary(
         packet, delta_shell, no_of_shells
     )
     assert packet.status == r_packet.PacketStatus.REABSORBED
@@ -483,7 +484,7 @@ def test_move_packet_across_shell_boundary_increment(
     energy = 0.9
     packet = r_packet.RPacket(r, mu, nu, energy)
     packet.current_shell_id = current_shell_id
-    r_packet.move_packet_across_shell_boundary(
+    tardis.montecarlo.montecarlo_numba.r_packet_transport.move_packet_across_shell_boundary(
         packet, delta_shell, no_of_shells
     )
     assert packet.current_shell_id == current_shell_id + delta_shell
@@ -641,7 +642,7 @@ def test_move_packet(
     numba_estimator = Estimators(
         packet_params["j"], packet_params["nu_bar"], 0, 0
     )
-    r_packet.move_r_packet(
+    tardis.montecarlo.montecarlo_numba.r_packet_transport.move_r_packet(
         packet, distance, time_explosion, numba_estimator
     )
 
@@ -809,7 +810,7 @@ def test_compute_distance2line_relativistic(
     distance = r_packet.calculate_distance_line(
         packet, comov_nu, nu_line, t_exp
     )
-    r_packet.move_r_packet(packet, distance, t_exp, numba_estimator)
+    tardis.montecarlo.montecarlo_numba.r_packet_transport.move_r_packet(packet, distance, t_exp, numba_estimator)
 
     doppler_factor = get_doppler_factor(r, mu, t_exp)
     comov_nu = packet.nu * doppler_factor
